Revised UpdateService API
Remove `succeeded` method. Assuming that updates are cheap and will be
done frequently, we now act as if updates succeeded and there were no
changes in the event of failure. There's no need to check for the
success of the operation now as it returns reasonable defaults (the last
ETag and current time).
This also allowed removal of the Optional types from the getters in
ModelUpdater.