Removed duplicate state from model
ETag fields were previously maintained in both ServiceManager and Model,
the latter for purposes of writing them to the cache. This change makes
it so ServiceManager is the only place they are kept, and Model cache
writes take explicit (or default) parameters instead.
Also reordered the arguments to cache-related methods and classes for
consistency.