Removed deferred issue loading
When loading from the cache, issue loading has to be deferred till after
all the other resources are loaded as it requires references to actual
objects in memory.
In order to realise this, issue loading was carried out in a later stage
of the pipeline (populateComponents instead of getReosurces), which broke
abstraction and was bad for code locality.
This has not been needed since the loading of other resources was changed
to become synchronous and is finally removed with this commit.