Propagated model update result back out
Now the model update will stop halfway if it detects that the repository
has changed midway through the update process. The cache won't be
written so there should no longer be data corruption.

This is intended as a safety measure. It should not happen if there are
no concurrency problems.