Lazy initialisation of issue label references
This resolves assertion failures caused by the changes to when issues
were loaded into the model. Previously issues would have to be loaded
last (and separately) as that would immediately resolve references to
related labels and milestones, and said resources would have to be
present. Now these references will only be loaded on first getter
access, which should be well after the model is completely initialised.