Refactored login process
Login is now done on a separate thread, to prevent the (admittedly
minor) UI freeze when it happens
There no longer is a progress dialog. The default repo will be
downloaded after the login dialog is closed. This way the login dialog
is only responsible for collecting initial input and saving it.