Added notion of pending repositories
This prevents multiple synchronous requests for the same repository from
downloading it multiple times (because the repository is only registered
as present upon completion of its download). This is due to the way the
UI is implemented, as panels request the repositories in their filter
boxes on every refresh in order to reflect changes to default
repository.