Changed reconciliation strategy for labels/milestones/users
Previously reconciliation of changes would be done by finding
corresponding entries in the existing resources and replacing those
with new ones.

This is bad for resources other than issues because they don't seem to
have an identity. Besides, they can't be selectively downloaded, and
there aren't a lot of them, so we use a much simpler strategy: simply
replacing the old ones with the new.