Added triggerMetadataUpdate to ModelUpdatedEvent
This controls whether or not to trigger metadata updates on refreshes on
the UI side.

We get metadata on refreshing, and refresh on getting
metadata. This leads to an infinite cycle which constantly gets
metadata. The data carried by the extra field helps us know when to
break the cycle.