Fixed bug read status was not propagated on update
The problem is that we set the the read status of issues in
MultiModel::add instead of MultiModel::addPending. The first happens on
every change to the model, including updates, whereas the second happens
only when a new model is added.

The fix is to move it to addPending: that's the only time we should
reset the read status of an issue to what is in Preferences.