Fixed bugs due to transient state not being transferred
Bugs:

- On refreshing a panel with an updated filter, all the events and
  comments would disappear before being loaded again
- After marking an issue as read, markedReadAt was cleared

The cause was that all transient state (metadata, read status) was not
being transferred when issues were updated. This was due to TurboIssues
being constructed from new external Issue objects that were created as
part of the update process.

The fix is to explicitly ensure that transient state is transferred.