Removed UI status message clear call after each applyfilterexpression
The message clear call made sense before the filter expressions caused additional data to be downloaded from the GitHub source (i.e. due to repo and/or updated filters). The message clear call arrives after the various status messages that signifying the download of additional data, which causes the status bar to be empty even though the program has not delivered all data to the UI yet.