Change bypasslogin credentials to test/test
Prior to this commit, when --bypasslogin=true is specified, the
username and password are not set, and are both empty strings when the
program is running. This is inconsistent with test mode behaviour when
--bypasslogin=true is not specified, in which only test/test are accepted
as credentials.

Both username and password are now set to "test" when --bypasslogin=true
is specified.