Add ADD_COMMENT as an UpdateDummyRepoEvent type
Comments are currently only relevant to us in terms of their authors and
creation dates, to display on the GUI and to process non-self update times
of issues. Thus, the actor field is introduced to UpdateDummyRepoEvent to
reflect the author of the commit, as well as to be used to specify the
actor of a model update in general, if needed in the future.

Since a new field is introduced, other constructors have to be modified to
accommodate the field.