Make update calls use correct constructor
Some calls in ModelUpdateUITest are calling the wrong constructor to
create UpdateDummyRepoEvents. The reason why this does not break anything
yet is that the switch in DummyRepo checks for the correct UpdateType
before calling a function, so the right functionality is still carried
out, but this should be fixed anyway.