Expand updateIssues to deal with nonself updates
Issue with ID 9 in dummy/dummy, now coming with a self-update when the
dummy repo state is initialized, should not show up in the issue list as
the update is by the HT user (which is "test"). Then, issue ID 9 is
updated with an non-self-update, causing it to show up in the issue list
with the "updated" filter.

The FutureTask to count issues shown in the panel is also refactored to
its own method as the same code is reused multiple times in the same test.