Generate UpdateDummyRepoEvent with factory methods
Previously, UpdateDummyRepoEvents of different types (e.g. to edit a new
issue vs creating a new issue) were created using overloaded
constructors. The problem is that these overloaded constructors do not
check if the UpdateType matches the purpose of the constructor, and assume
that they are correct based on the number of arguments given.

This is dangerous if the constructor used does not match the UpdateType.
For example, a new UpdateDummyRepoEvent can be called with the UpdateType
UPDATE_MILESTONE with the idString field filled out instead of the itemId
field. This causes the milestone with ID -1 to be updated, as per the
default value when the idString field is provided.

To fix this, instead of using overloaded constructors, each UpdateType now
has a corresponding static factory method that creates the appropriate
UpdateDummyRepoEvent.