Fixed logic error in label matching
The problem was returning false in the loop. This could abort the
qualifier's checks early, even if there was a label that came later that
would satisfy the qualifier.

This caused some issues to be wrongly rejected (or included, if negation
was used).

Now the qualifier only rejects an issue if all labels have been checked.
There's a test for this specific case. The logic for matching labels is
factored out so this sort of thing cannot happen again.