Give metadata from separate DummyRepoState hashset
To match the new metadata reconciliation logic using ETag, which only
replaces metadata if the issue ETag is different, we need a mechanism to
return a different ETag only if there were updates to the particular issue
being retrieved.

Thus, we maintain the HashSet updatedMetadata which contains issues that
will have a different ETag the next time metadata is retrieved i.e. issues
that have been modified since the last time metadata was retrieved. In
this manner, after metadata is retrieved, the issueId is removed from the
updatedMetadata HashSet until the issue is modified again.

We also extract IssueMetadata from issues and store them in the HashMap
issueMetadata. This is to ensure that getUpdatedIssues does not retrieve
the ETag prematurely - instead, issue ETags should only be retrieved
through a getUpdatedEvents call.