Make modifications update updatedMetadata
Continuing from the previous commit, instead of adding the updated
metadata to the issue itself, we insert the metadata into the map
issueMetadata and update updatedMetadata for the three methods that modify
an issue: commentOnIssue, setLabels and updateIssue.

Also, for each of these methods, the issues/metadata being mutated to
reflect modification are copies of the original objects in DummyRepoState,
which are inserted into the issues, updatedIssues, issueMetadata maps to
replace the originals. This is so that the changes do not get reflected
immediately in the GUI, and only are shown when a refresh or metadata
retrieval is triggered, which is the normal behaviour of the program.

The above copying and re-inserting of issues/metadata during mutation are
refactored into the auxiliary methods produceMutables and markUpdatedIssue
to ensure DRY.