Modify test mode API quota behaviour
The test mode API quota now decreases whenever a getUpdatedEvent call is
carried out for an issue with new updates since last metadata retrieval,
and whenever getComments is called. When getUpdatedEvent is called for an
issue that does not have new updates, the API quota will not decrease, as
per the GitHub API behaviour.

Other method calls such as getIssues or getMilestones also do not change
the API quota.

Additionally, the API box is assigned an ID so that tests can locate the
API box and its text easily.