Pass TurboIssues instead of ETag HashMaps
While TurboIssues are not fully immutable, IssueMetadata is immutable and
thus instead of converting the TurboIssues to be updated into a HashMap of
their IDs and ETags, we can simply pass TurboIssues to the backend, where
the ETags and IDs can be extracted and dispatched as metadata updates.

This commit mainly modifies the GUIController populateUpdateList method,
which tallies up issues to update in the form of ArrayList of TurboIssues,
instead of the HashMap of their IDs and ETags. Also, as the object type of
issuesToUpdate is now modified, the method signatures of the call chain
from Logic to DownloadMetadataTask have to be modified to fit the new
issuesToUpdate.

At DownloadMetadataTask, we also do an extra step to extract the ETags and
IDs from the list of issues. This step used to be carried out in the
GUIController, but it is now passed down to DownloadMetadataTask as
GUIController is sending TurboIssues and not doing the extraction.