Incorporate deepCopy for test resource retrieval
This commit makes use of the deepCopyIssues/Milestones/Labels/Users
methods in passing repository information stored within DummyRepoState to
the rest of the application. This thus ensures that the outgoing objects
can be freely mutated without any effect on the source-of-truth objects
inside DummyRepoState unless an API call is sent to explicitly modify the
source-of-truth.

Add copy constructors to TurboLabel and TurboUser

Only TurboIssues and TurboMilestones have copy constructors. In order to
facilitate deep copying, copy constructors for TurboLabel and TurboUser
have also been added.

Add deep copy methods for DummyRepoState resources

The DummyRepoState, when queried from the application, returns
source-of-truth objects. However, this means that when the application
mutates the objects returned by DummyRepoState, the source-of-truth
objects as held by the DummyRepoStates are also affected.

While the objects held by the application may be mutated for various
reasons e.g. changing the look of the UI, this action should not affect
the source-of-truth as long as we are not making any API calls to the
repo source. DummyRepoState thus is not exerting this behaviour. This
commit changes the objects that DummyRepoState returns from
source-of-truth objects to only copies of the source-of-truth objects,
and thus solves the problem.

It should be noted that while resource objects such as TurboIssues and
TurboLabels are currently mutable, they should be made immutable in the
future, as per discussions in #849. Once the resource objects have been
made mutable, the copying process, as implemented by this commit, will no
longer be necessary.