Parse events in GitHubEventsResponse constructor
The parseEventParameters call was deleted with the intention of removing
its guard clause checking if it is null, as we now use a NullInputStream
object instead of the null value, and thus we don't have to avoid
NullPointerExceptions using the guard clause any more.

parseEventParameters parses the JSON of the response from GitHub into
TurboIssueEvent objects, and thus the events in the response will be empty
if it is not called.

As parseEventParameters is only called from the constructor to modify the
turboIssueEvents object field, this commit also modifies
parseEventParameters' method signature to make GitHubEventsResponse
immutable.