Remove focus from FilterTextField upon init
So that upon startup the focus is on the panel and not on the
FilterTextField of any panel, and the navigation shortcut works at
startup.
Did the same for opening a Board for consistency.
Also fix a bug where multiple panels can have focused colour. This is due to using original index instead of current index for click event.
Some refactoring of panel scroll.