Add GitHubResponsesTest
This tests the code in GitHubEventsResponse responsible for parsing GitHub
API response bodies into TurboIssueEvent objects. At the time of commit,
coverage of GitHubEventsResponse is not required (and will not count
towards total coverage), but this test will prevent regression in case
GitHubEventsResponse code is changed.

The method gitHubEventsResponseJsonStream produces an InputStream
containing data in the same format as actual GitHub responses, albeit
stripped down to the bare minimum required fields for the functionalities
of HubTurbo. In the event of the GitHub API changing format, the string
within this method's body should also be modified accordingly.

GitHubResponsesTest should, in the future, be expanded to work with not
just events responses but also other resources' responses such as within
UpdateService's logic for paging.