Add commentsETag to IssueMetadata
The ETags for comments and events for an issue are maintained separately
by GitHub. If we keep track of only the events ETag, no new comments will
be downloaded until an update has been done to the event e.g. relabel,
rename, remilestone.

This commit adds support for storing and keeping track of updated comments
separately in the test stub, as well as create the commentsETag field in
IssueMetadata. This commit also changes the MultiModel state update logic
to also take into account commentsETag and not just the events ETag, but
does not yet implement logic in the GitHubRepo backend to store and attach
ETags while retrieving comments.