Backported fixes from TestFX 3.1.2
TestFX 3.1.2 introduces methods for pressing keys without waiting on the
JavaFX Application Thread and and for pressing whole key combinations.
Unfortunately that version also introduces regressions and is no longer
maintained, with 4.x in alpha.

This commit implements the useful methods to fix an issue with tests,
where pressing keys can cause any number of key-press events to fire
depending on timing.