Download comments regardless of stored ETags
Until comment ETags are implemented, we download comments regardless of
whether the events ETags returned are similar.

The setMetadata logic in MultiModel has to be modified so that we have
more granularity in reconciling metadata i.e. we have the choice to
avoid replacing events (in case of no new events, for example) while
keeping the comments of the issue intact. This is implemented by moving
the ETag comparison guard clause in MultiModel to the TurboIssue
constructor called in the guarded block.

Proper implementation of comment ETags have to be done in a different PR,
as it is necessary to manually send and parse the requests and responses
without the help of egit if we want to make use of ETags in the requests
and responses.