Overloaded `press` for KeyCodeCombinations
This is the same as chaining a sequence of presses, then releases.
pushKeys' lack of delay doesn't seem to be the solution in all cases,
so we still need press' semantics.

Also added manual conversion of Shortcut keys to either Ctrl or Meta
depending on platform. Shortcut works fine for users, but isn't defined
for the AWT robot used by TestFX to push keys.