Fixed nondeterminism in PlatformExTest
We are not actually guaranteed that the runnable given to
Platform.runLater will execute (on the UI thread) after what immediately
follows it (which executes on an arbitrary thread), though that is what
usually happens.

This gets rid of that assumption. The only thing we can say for sure
about about the result there is that it should be 1 after the
invocation of waitOnFxThread.