Add filter and sort functionality to Logic
This commit implements the full filter and sort logic within Logic, which
is the same as the filter and sort functionality in GUIController (soon to
be removed).

filterSortRefresh is the logical equivalent of processPanel in the
GUIController. It tallies issues for a metadata update (first filter), and
then after the updates are done, it filters and sorts issues to be
displayed to the user.

refreshUI calls filterSortRefresh on all GUI filters. It does this through
a call of getAllUIFilters.

getAllUIFilters is a way for Logic to retrieve filter epxressions from the
GUI. The reason why Logic now actively retrieves data from the GUI (in the
form of filters) is that filtering and sorting is done after opening a
repository, and it does not make sense for an openRepository call to
originate from the GUIController (from which we can pass the current
filters to Logic). This also applies, though to a lesser extent, to a
refresh triggered by the TickingTimer or an F5 input by the user. What
this means is that passing filter expressions from the GUIController to
Logic as function arguments is not viable. Thus, Logic has to actively
retrieve FilterExpressions from the GUI.

openRepositoriesInFilters is the logical equivalent of
openAllReposInExpression in GUIController. It dispatches open repository
requests to repoIO asynchronously.

getIssueMetadata is stripped of the updateUIAndShow call after finishing
metadata retrieval, as well as the API limits refresh call.

updateUI is modified to carry the list of issues to be displayed by the
GUI, instead of a hasMetadata flag (this flag is split into multiple flags
contained within the list of issues).

determineComparator is copied completely from GUIController. However, with
the current implementation, it is possible to remove the
hasUpdatedQualifier flag in its method signature later.

tallyMetadataUpdate and filterAndSort contain the logic for the first and
second filters respectively.