Remove filter and sort from GUIController
Most methods within GUIController is no longer necessary, as they have a
counterpart in Logic. processPanel, dispatchMetadataUpdates,
filterAndSortPanel, determineComparator, populateUpdateList are all
removed.

The logic of modelUpdated is modified to reflect the change in structure
of ModelUpdatedEvent. For each active panel, the list of issues carried by
ModelUpdatedEvent is searched for a matching filter expression, and if
found, the panel is updated with the corresponding issues.

panelFilterExpressionChanged logic is modified so that instead of calling
processPanel, it calls filterSortRefresh in Logic, passing in only its own
filter expression as an argument (as opposed to all panels' filter
expressions)

The multimodel stored in the GUIController is also removed, as
panelFilterExpressionChanged logic is now modified so that it no longer
requires a multimodel living in the GUIController.