Board Auto-create - Milestone
Added Auto-create option into Board. This commit adds ability to auto-create milestone board.
This commit removes milestone alias clamping. If an alias is out of range, it will be parsed as it is, e.g. `curr+100` will be regarded that user is looking for milestone named curr+100.

Introduced in this commit:
- Utility.generateName which would generate name that does not overwritten any existing names.
- DialogMessage.showInformationDialog.
- Qualifier.FALSE. It will fail for all issue; currently used to denote alias does not work
- waitUntilNodeAppear/Disappear for Matcher<>; for consistency in UITest