Migrate filter and sort to UpdateController
For clearer code. The core getIssueMetadata still lies in Logic as it
requires access to RepoIO. For this, updateRateLimits and updateUI has to
be made public so that UpdateController can call them.

Filtering and sorting also requires access to models, hence it is exposed
to UpdateController through getModels.

GUIController cannot call filterSortRefresh directly anymore (when it
wants to process a single panel's filter). The new refreshPanel method
bridges this gap.