Create GUIElement class
Each GUIElement contains references to a TurboIssue, as well as related
objects which are necessary to properly the TurboIssue as a panel card.
These include:
- Labels related to the issue (both current and removed)
- Milestone of the issue (if exists)
- Assignee of the issue (if exists)

Removed labels of the issue are also necessary to display the issue
metadata, in particular label removal events.

GUIElements are meant to ultimately remove the references to the Model by
instead passing only the references to the necessary objects within the
Model. This means that together with model deletion, obsolete objects
within a Model can be quickly (and safely) garbage collected.