UpdateController filtering produces GUIElements
Instead of producing a filtered and sorted list of TurboIssues and passing
it to the GUIController to be displayed, UpdateController now contains
logic to search for objects relevant to each TurboIssue from the
MultiModel, and then wraps them together in a GUIElement.

The logic to search for relevant objects is taken from the MultiModel, and
are similarly named.