Modify ModelUpdatedEvent fields
Added elementsToShow, which replaces issuesToShow.
Added current list of users relevant to the program. Originally retrieved
by FilterPanel directly from the MultiModel in its ModelUpdatedEvent
handling code (onModelUpdate), so as to add user IDs as keywords for the
filter autocomplete.
Added the defaultRepoId. Originally retrieved directly from the MultiModel
in various places in the GUI to assign IDs to nodes (e.g.
hubturbo/hubturbo_col1_...).

Removed issuesToShow (TurboIssues are already encapsulated in
elementsToShow), as well as the reference to the MultiModel.