Display ListPanelCards using GUIElements
Switch out references to the model with method calls on GUIElements to
display ListPanelCards.

This involves changing the type of the IssueListView from
NavigableListView<TurboIssue> to NavigableListView<GUIElement>.

The FilterPanel, instead of holding an ObservableList of TurboIssues, now
holds a corresponding ObservableList of GUIElements.

ListPanelCard now takes in only a relevant GUIElement, instead of the
MultiModel and a TurboIssue. The rendering code stays the same, as only
one extra step is taken to extract the TurboIssue out of the GUIElement,
instead of interacting immediately with the TurboIssue.

ListPanelCell thus does not need a reference to the MultiModel any more.
Both code in ListPanelCell and ListPanelCard undergo one extra step to
extract the TurboIssue from the currently selected item in IssueListView,
but the code is otherwise unchanged.

TurboIssueEvent is modified to extract relevant information to produce the
Node object from the GUIElement instead of the entire MultiModel. This is
in turn called by ListPanelCard, under layoutEvents.