Make GUIController source of defaultRepoID in GUI
Instead of retrieving the current default repo's ID directly from the
MultiModel, classes in the GUI that use the default repo's ID to assign
IDs to various Nodes now retrieve this information from the GUIController.

The GUIController, in turn, updates this information whenever a
ModelUpdatedEvent is triggered. This ensures that the default repo's ID is
synchronised between the MultiModel and the GUI, as a successful switch of
the default repo will trigger a ModelUpdatedEvent via Logic.updateUI.