Pass defaultRepoId to GUIController with event
Instead of letting ModelUpdatedEvent carry the ID of the default
repository to the GUIController, we have GUIController react to a
PrimaryRepoChangedEvent. This is similar to how the GUIController updates
its API rate limit box.

The ModelUpdatedEvent cannot carry the defaultRepoId to the GUIController
due to the fact that panels.init call in UI upon startup runs before the
openRepository in Logic finishes, which results in filter text fields and
buttons on panels being set with defaultRepoIds before the first
ModelUpdatedEvent is fired from Logic, resulting in IDs like
null_col0_filterTextField.

Reverted behaviour of filter text field changing its own ID in response to
a change in the default repo, as buttons and the panel name text do not
change their IDs, and there is no simple mechanism available to make these
nodes adapt their IDs to a backend change in the default repository (other
than events, but smaller GUI classes handling events can be messy).