Reconcile labels in DummyRepoState.setLabels
Currently, setLabels would add unlabeling events for all current labels,
and then add labeling events for all new labels. This does not reflect
actual GitHub behaviour, which only unlabels those not in the new list of
labels, and adds only those not in the current list of labels. This commit
changes the behaviour of setLabels to match that of the GitHub API.